webapp refresh commit

+++ src/main/webapp/member/deletePro.jsp
@@ -13,8 +13,15 @@
 
 <%
     request.setCharacterEncoding("utf-8");
-    %>
+%>
 
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
 <%
 	String passwd = request.getParameter("passwd");
 
@@ -22,6 +29,12 @@
 	if(id==null || id.equals("")) {
 		response.sendRedirect("../main/main.jsp");	
 	} else {
+		FreeboardDAO fdao = FreeboardDAO.getInstance();
+		fdao.deleteMemberFreeboard(id);
+		
+	
+		ReplyDAO rdao = ReplyDAO.getInstance();
+		rdao.deleteMemberreply(id);
 		
 		//페이지처리
 		int pageSize = 20;
@@ -36,35 +49,43 @@
 		int startRow = (currentPage - 1) * pageSize + 1;//(1-1)*10+1=1	/(2-1)*10+1=11
 		int endRow = currentPage * pageSize;			//1*10=10		/(2*10)=20
 		
-		List<ReplyVO> rlist = null;
-		int rcount = 0; //총 댓글수	
-		
-		ReplyDAO rdao = ReplyDAO.getInstance();
-		rcount = rdao.getArticleCount();
-		rdao.deleteMemberreply(id);
-		
-		rlist = rdao.getMemberArticles(id, startRow, pageSize);
-		
-		int[] ref;
-		ref = new int[rlist.size()];
-		
-        for (int i = 0; i < rlist.size(); i++) {
-    		ReplyVO reply = rlist.get(i);
-        	ref[i] = reply.getRef();
-        }
-		
-		
+		List<QnaboardVO> qlist = null;
 		
 		QnaboardDAO qdao = QnaboardDAO.getInstance();
-        for (int i = 0; i < rlist.size(); i++) {
-    		qdao.deleteqnaboard(ref[i]);
-        }
+		qlist = qdao.getMemberArticles(id, startRow, pageSize);
+		
+		if(qlist != null){
+			int[] ref;
+			ref = new int[qlist.size()];
+			
+	       for (int i = 0; i < qlist.size(); i++) {
+	       	QnaboardVO qnareply = qlist.get(i);
+	       	ref[i] = qnareply.getNum();
+	       }
+		
+	       for (int i = 0; i < qlist.size(); i++) {
+	   		qdao.deleteqnaboard(ref[i]);
+	       }
+		}
 		
 		MemberDAO mdao = MemberDAO.getInstance();
-		mdao.deleteMember(id, passwd);			
-		
-		session.invalidate();
-		out.println("<script>alert('회원 탈퇴되었습니다.');</script>");
-		out.println("<script>document.location.href='../main/main.jsp';</script>");
+		int check = mdao.deleteMember(id, passwd);
+		if (check == 1){
+			session.invalidate();
+			out.println("<script>alert('회원 탈퇴되었습니다.');</script>");
+			out.println("<script>document.location.href='../main/main.jsp';</script>");
+		}
+		else{
+	%>
+		<script type="text/javascript">
+			alert("비밀번호가 맞지 않습니다.");
+			history.go(-1);
+			document.delete_form.passwd.focus();
+		</script>
+<%
+		}
 	}
 %>	
+
+</body>
+</html>
